Allow nil current temperature in C
Thanks @richard-otto-obrien for this, spotted in https://github.com/ccutrer/balboa_worldwide_app/pull/20

If current temperature is not yet known, we store it as `nil`,
but in C mode we then attempt to divide this by 2. Don't do that.
For consistency, treat set and current temperature the same,
although at present set temperature can never be `nil`.