Automatically restart if bridge hangs
Sometimes corrupt data causes the data-processing thread to quit,
but the bridge keeps running. This means data is no longer transmitted
to the MQTT server, but systemd does not notice the problem.

This enhancement adds systemd support to `bwa_mqtt_bridge`, such that
the app notifies systemd when it starts successfully, and every time
it completes processing a status message. It then configures systemd
so that it restarts the app if it isn't notified for a while.

These changes have no effect if systemd is not in use (specifically,
if `NOTIFY_SOCKET` is not defined in the environment).