add soak command
closes #63

which sets all pumps to speed 0, regardless of current speed.
use it to set to speed 0 if possible (to avoid having to cycle
through multiple speeds). only applicable if you have a single
pump, and that pump has multiple speeds. you can still send soak
explicitly with the command property